:sparkles: script to remove unwanted methods
